% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic-model.r
\name{add_generic_positions_constraint}
\alias{add_generic_positions_constraint}
\title{Add custom position constraints}
\usage{
add_generic_positions_constraint(model, data, constraints)
}
\arguments{
\item{model}{a coach model}

\item{data}{data frame with fpts projections}

\item{constraints}{a named list with constraints. Flex or Wildcard positions should be named with a / between positions. See examples.}
}
\description{
Add custom position constraints
}
\examples{
\dontrun{
data <- read_dk("yourfilenamehere.csv")
constraints <- list("QB" = 1, "RB" = 2, "WR" = 3, "TE" = 1, "RB/WR/TE" = 1, "DST" = 1)
model <- model_generic(data, total_salary = 50E3, roster_size = 9, max_from_team = 4)
model <- add_generic_positions_constraint(model, data, constraints)
}
}
