% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlb.r
\name{add_pitcher_hitter_constraint}
\alias{add_pitcher_hitter_constraint}
\title{Don't let us build lineups with pitchers and hitters facing each other}
\usage{
add_pitcher_hitter_constraint(model, mlb)
}
\arguments{
\item{model}{ompr model}

\item{mlb}{data frame of mlb data}
}
\description{
Don't let us build lineups with pitchers and hitters facing each other
}
\keyword{internal}
