% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-dk.r
\name{add_row_id}
\alias{add_row_id}
\title{Add a row id column to the data frame}
\usage{
add_row_id(df)
}
\arguments{
\item{df}{data frame}
}
\description{
Add a row id column to the data frame
}
\keyword{internal}
