% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic-model.r
\name{add_stack_size_constraint}
\alias{add_stack_size_constraint}
\title{Stack best available teams with variable sizes}
\usage{
add_stack_size_constraint(model, mlb, stack_sizes, stack_teams = NULL)
}
\arguments{
\item{model}{model}

\item{mlb}{data}

\item{stack_sizes}{size of each stack}

\item{stack_teams}{subset of teams to select for stacking. leave NULL to allow any team}
}
\description{
Stack best available teams with variable sizes
}
\keyword{internal}
