% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asserts.r
\name{assert_coltypes}
\alias{assert_coltypes}
\title{Make sure columns in data frame are the right type}
\usage{
assert_coltypes(df, cols, coltypes)
}
\arguments{
\item{df}{data frame}

\item{cols}{column names}

\item{coltypes}{column types (ex. "integer", "double", "character")}
}
\description{
Make sure columns in data frame are the right type
}
\keyword{internal}
