% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asserts.r
\name{assert_has_cols}
\alias{assert_has_cols}
\title{Assert that the dataframe has certain columns}
\usage{
assert_has_cols(df, cols)
}
\arguments{
\item{df}{a data frame}

\item{cols}{column names}
}
\description{
Assert that the dataframe has certain columns
}
