% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draftkings-mlb.r, R/draftkings-nba.r,
%   R/draftkings-nfl.r, R/fanduel-mlb.r, R/fanduel-nba.r, R/fanduel-nhl.r,
%   R/fanduel-wnba.r, R/fantasydraft-mlb.r, R/fantasydraft-nba.r,
%   R/fantasydraft-nfl.r, R/generic-model.r
\name{model_dk_mlb}
\alias{model_dk_mlb}
\alias{model_dk_nba}
\alias{model_dk_nfl}
\alias{model_fduel_mlb}
\alias{model_fduel_nba}
\alias{model_fduel_nhl}
\alias{model_fduel_wnba}
\alias{model_fdr_mlb}
\alias{model_fdr_nba}
\alias{model_fdr_nfl}
\alias{model_generic}
\title{Generic fantasy model}
\usage{
model_dk_mlb(data, existing_rosters = list())

model_dk_nba(data, existing_rosters = list())

model_dk_nfl(data, existing_rosters = list())

model_fduel_mlb(data, existing_rosters = list())

model_fduel_nba(data, existing_rosters = list())

model_fduel_nhl(data, existing_rosters = list())

model_fduel_wnba(data, existing_rosters = list())

model_fdr_mlb(data, existing_rosters = list())

model_fdr_nba(data, existing_rosters = list())

model_fdr_nfl(data, existing_rosters = list())

model_generic(data, total_salary, roster_size, max_from_team = 4,
  existing_rosters = list())
}
\arguments{
\item{data}{data frame}

\item{existing_rosters}{list of integer vectors specifying rosters to exclude}

\item{total_salary}{maximum salary available for each lineup}

\item{roster_size}{number of players on each lineup}

\item{max_from_team}{maximum number of players allowed from each team}
}
\description{
Generic fantasy model
}
