% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.r
\name{normalize_lineup}
\alias{normalize_lineup}
\title{Normalize a lineup based on position}
\usage{
normalize_lineup(lineup, site = c("draftkings", "fanduel",
  "fantasydraft"), sport = c("nfl", "mlb", "nba", "nhl", "nascar"),
  colname = "position")
}
\arguments{
\item{lineup}{a data frame lineup}

\item{site}{string}

\item{sport}{string}

\item{colname}{default column to apply normalization to}
}
\description{
Applies a utility or flex tag to appropriate lineup positions
}
