% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic-optimize.r
\name{optimize_generic}
\alias{optimize_generic}
\title{Optimize a fantasy Model}
\usage{
optimize_generic(data, model, L = 3L, solver = c("glpk", "symphony",
  "cbc"), bans = NULL, locks = NULL, stack_sizes = NULL,
  stack_teams = NULL, min_salary = NULL, max_exposure = 1)
}
\arguments{
\item{data}{projected fantasy points}

\item{model}{optimization model}

\item{L}{total number of lineups}

\item{solver}{ROI solver to use}

\item{bans}{row_ids of players to exclude from all lineups}

\item{locks}{row_ids or players to include in all lineups}

\item{stack_sizes}{size of each stack}

\item{stack_teams}{subset of teams to use to generate stacks. NULL will use all teams.}

\item{min_salary}{minimum salary to use}

\item{max_exposure}{max exposure for all players or a vector of exposures for each player}
}
\description{
Optimize a fantasy Model
}
