% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic-optimize.r
\name{optimize_generic_one}
\alias{optimize_generic_one}
\title{Optimize fantasy lineups}
\usage{
optimize_generic_one(data, model, solver = c("glpk", "symphony", "cbc"))
}
\description{
Optimize fantasy lineups
}
\keyword{internal}
