% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{parse_constraints}
\alias{parse_constraints}
\title{Parse custom constraints}
\usage{
parse_constraints(x)
}
\arguments{
\item{x}{a list of constraints. See examples}
}
\description{
Parse custom constraints
}
\examples{
\dontrun{
x <- list("QB" = 1, "RB" = 2, "WR" = 3, "TE" = 1, "RB/WR/TE" = 1, "DST" = 1)
parse_constraints(x)
}
}
\keyword{internal}
