% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{parse_locations}
\alias{parse_locations}
\title{Parse locations from a string like SAS@GSW}
\usage{
parse_locations(x)
}
\arguments{
\item{x}{vector of strings}
}
\description{
Parse locations from a string like SAS@GSW
}
