% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-dk.r
\name{read_dk}
\alias{read_dk}
\title{Read Draftkings}
\usage{
read_dk(path, colnums)
}
\arguments{
\item{path}{path to csv file}

\item{colnums}{columns that contain the roster data}
}
\description{
Reads in raw file and then cleans it up
}
\keyword{internal}
