% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.r
\name{read_dk_nfl}
\alias{read_dk_nfl}
\alias{read_dk_mlb}
\title{Read Draftkings Player Pool}
\usage{
read_dk_nfl(path)

read_dk_mlb(path)
}
\arguments{
\item{path}{path to csv file}
}
\description{
Expects a csv type from \url{https://www.draftkings.com/lineup/upload}
}
