% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-fd.r
\name{read_fd}
\alias{read_fd}
\title{Read Fanduel Players}
\usage{
read_fd(path)
}
\arguments{
\item{path}{path to csv file}
}
\description{
Expects a csv from \url{https://www.fanduel.com/} obtained by clicking
"Download Players List" for your contest page.
}
