% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{where}
\alias{where}
\title{Same as which, but returns NULL instead of empty vector when no match}
\usage{
where(x)
}
\description{
Same as which, but returns NULL instead of empty vector when no match
}
\keyword{internal}
