% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.r
\name{convert_lineup}
\alias{convert_lineup}
\title{Convert lineup to submission format}
\usage{
convert_lineup(lineup)
}
\arguments{
\item{lineup}{a normalized lineup}
}
\description{
Convert lineup to submission format
}
\keyword{internal}
