% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.r
\name{convert_lineup}
\alias{convert_lineup}
\title{Convert lineup to submission format}
\usage{
convert_lineup(lineup, site = c("draftkings", "fanduel", "fantasydraft"),
  sport = c("nfl", "mlb", "nba", "nhl", "nascar"), ...)
}
\arguments{
\item{lineup}{a normalized lineup}

\item{site}{name of DFS site}

\item{sport}{name of sport}

\item{...}{additional arguments passed to \code{\link{normalize_lineup}}}
}
\description{
Convert lineup to submission format
}
\keyword{internal}
