% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draftkings-mlb.r, R/draftkings-nascar.r,
%   R/draftkings-nba.r, R/draftkings-nfl.r, R/draftkings-nhl.r, R/fanduel-mlb.r,
%   R/fanduel-nascar.r, R/fanduel-nba.r, R/fanduel-nfl.r, R/fanduel-nhl.r,
%   R/fanduel-wnba.r, R/fantasydraft-mlb.r, R/fantasydraft-nba.r,
%   R/fantasydraft-nfl.r, R/generic-model.r
\name{model_dk_mlb}
\alias{model_dk_mlb}
\alias{model_dk_nascar}
\alias{model_dk_nba}
\alias{model_dk_nfl}
\alias{model_dk_nhl}
\alias{model_fd_mlb}
\alias{model_fd_nascar}
\alias{model_fd_nba}
\alias{model_fd_nfl}
\alias{model_fd_nhl}
\alias{model_fd_wnba}
\alias{model_fdr_mlb}
\alias{model_fdr_nba}
\alias{model_fdr_nfl}
\alias{model_generic}
\title{Generic fantasy model}
\usage{
model_dk_mlb(data, existing_rosters = list())

model_dk_nascar(data, ...)

model_dk_nba(data, ...)

model_dk_nfl(data, existing_rosters = list())

model_dk_nhl(data, ...)

model_fd_mlb(data, existing_rosters = list())

model_fd_nascar(data, ...)

model_fd_nba(data, ...)

model_fd_nfl(data, existing_rosters = list())

model_fd_nhl(data, existing_rosters = list())

model_fd_wnba(data, ...)

model_fdr_mlb(data, existing_rosters = list())

model_fdr_nba(data, ...)

model_fdr_nfl(data, existing_rosters = list())

model_generic(data, total_salary, roster_size, max_from_team = 4,
  existing_rosters = list(), type = c("binary", "continuous"))
}
\arguments{
\item{data}{data frame}

\item{existing_rosters}{list of integer vectors specifying rosters to exclude}

\item{...}{extra arguments passed to \code{\link{model_generic}}}

\item{total_salary}{maximum salary available for each lineup}

\item{roster_size}{number of players on each lineup}

\item{max_from_team}{maximum number of players allowed from each team}

\item{type}{binary or continuous optimization}
}
\description{
Generic fantasy model
}
