% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.r
\name{normalize_positions}
\alias{normalize_positions}
\alias{normalize_dk_nfl}
\alias{normalize_dk_nba}
\alias{normalize_dk_nhl}
\alias{normalize_fd_nfl}
\alias{normalize_fd_mlb}
\alias{normalize_fd_nascar}
\alias{normalize_fdr_nba}
\title{Normalize positions by allowed counts}
\usage{
normalize_positions(pos, pos_max, wildcard)

normalize_dk_nfl(pos)

normalize_dk_nba(pos)

normalize_dk_nhl(pos)

normalize_fd_nfl(pos)

normalize_fd_mlb(pos)

normalize_fd_nascar(pos)

normalize_fdr_nba(pos)
}
\arguments{
\item{pos}{a character vector of existing positions}

\item{pos_max}{a named numeric vector giving the max allowed players at that position}

\item{wildcard}{a string specifying what to replace positions with}
}
\description{
Will replace positions with wildcard if their count exceeds the max allowed.
This is commonly used to set a player to a utility or flex position.
}
\examples{
\dontrun{
 pos <- c("P", "1B", "1B", "2B", "3B", "SS", "OF", "OF", "OF")
 pos_max <- c("P" = 1, "1B" = 1, "2B" = 1, "3B" = 1, "SS" = 1, "OF" = 3)
 wildcard <- "UTIL"

 normalize_positions(pos, pos_max, wildcard)
}
}
\keyword{internal}
