% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.r
\name{read_dk}
\alias{read_dk}
\title{Read Draftkings}
\usage{
read_dk(path)
}
\arguments{
\item{path}{path to csv file}
}
\description{
Expects a csv from \url{https://www.draftkings.com/} obtained by clicking
"Export to csv" for your contest page.
}
