% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.r
\name{read_dk_raw}
\alias{read_dk_raw}
\title{Read Draftkings Players}
\usage{
read_dk_raw(path, colnums)
}
\arguments{
\item{path}{path to csv file}

\item{colnums}{columns that contain the roster data}
}
\description{
Expects a csv from \url{https://www.draftkings.com/lineup/upload}
}
\keyword{internal}
